+++ activiti-cloud-service-common/activiti-cloud-services-common-security-keycloak/src/main/java/org/activiti/cloud/services/common/security/keycloak/JwtAdapter.java~AAE-7798 remove keycloak adapter wip
