+++ activiti-cloud-service-common/activiti-cloud-services-common-security-keycloak/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.activiti.cloud.services.common.security.keycloak.config.CommonSecurityAutoConfiguration
+org.activiti.cloud.services.common.security.config.CommonSecurityAutoConfiguration
