+++ activiti-cloud-service-common/activiti-cloud-services-common-security-keycloak/src/test/java/org/activiti/cloud/services/common/security/keycloak/test/CommonSecurityAutoConfigurationIT.java
@@ -17,7 +17,7 @@ package org.activiti.cloud.services.common.security.keycloak.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.activiti.cloud.services.common.security.keycloak.config.CommonSecurityAutoConfiguration;
+import org.activiti.cloud.services.common.security.config.CommonSecurityAutoConfiguration;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringBootConfiguration;
