+++ activiti-cloud-service-common/activiti-cloud-services-common-security-keycloak/src/test/java/org/activiti/cloud/services/common/security/keycloak/test/JwtAccessTokenValidatorTest.java
@@ -23,9 +23,9 @@ import static org.mockito.Mockito.when;
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
-import org.activiti.cloud.services.common.security.keycloak.JwtAccessTokenValidator;
-import org.activiti.cloud.services.common.security.keycloak.config.JwtAdapter;
-import org.activiti.cloud.services.common.security.keycloak.config.KeycloakJwtAdapter;
+import org.activiti.cloud.services.common.security.jwt.JwtAccessTokenValidator;
+import org.activiti.cloud.services.common.security.jwt.JwtAdapter;
+import org.activiti.cloud.services.common.security.keycloak.KeycloakJwtAdapter;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
